removes duplicated tests.
